Keep traversals mainly in the traditional representation
This has several benefits:
- No need to duplicate the derivation code between the -optics and -lens.
   There might have been another way to avoid the duplication, but
   this is way easier.
- No need to implement a normalisation step for profunctor traversals.
   I experimented with going from Profunctor to VL, then using the
   'confusing' normalisation, then back to Profunctor, but the compile
   time was horrible.
- Writing VL traversals is actually quite natural.

A disadvantage is that the interface is somewhat inconsistent, but
that's a relatively small price to pay until there's a better solution.
