--- generic-lens-core/src/Data/Generics/Product/Param.hs
@@ -1,4 +1,0 @@
--- | 
-
-module Data.Generics.Product.Param where
-
